+++ src/experimentation-opportunities/handler.js
@@ -16,7 +16,6 @@ import { defaultProvider } from '@aws-sdk/credential-provider-node';
 import { GetObjectCommand, HeadObjectCommand } from '@aws-sdk/client-s3';
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
 import { AuditBuilder } from '../common/audit-builder.js';
-import { getRUMDomainkey } from '../support/utils.js';
 import s3Client from '../support/s3-client.js';
 import { wwwUrlResolver } from '../common/audit.js';
 import { convertToOpportunity } from '../common/opportunity.js';
@@ -293,10 +292,8 @@ export async function handler(auditUrl, context, site) {
   const { log } = context;
 
   const rumAPIClient = RUMAPIClient.createFrom(context);
-  const domainkey = await getRUMDomainkey(site.getBaseURL(), context);
   const options = {
     domain: auditUrl,
-    domainkey,
     interval: DAYS,
     granularity: 'hourly',
   };